It tests sliceByTime, provides a docstring for it, and the function is in the helper file instead of in chat.js
